services/processing/requirements.txt
flask
kafka-python
python-dotenv
prometheus-client
psycopg2-binary
